// front-end/packages/shared/src/models/bookLanguage.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 * OpenAPI documentation.
 * OpenAPI spec version: 2.0
 */

export type BookLanguage = (typeof BookLanguage)[keyof typeof BookLanguage];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookLanguage = {
  VN: "VN",
  EN: "EN",
  CN: "CN",
  JP: "JP",
} as const;