// front-end/packages/shared/src/models/bookType.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 * OpenAPI documentation.
 * OpenAPI spec version: 2.0
 */

export type BookType = (typeof BookType)[keyof typeof BookType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookType = {
  HARD_COVER: "HARD_COVER",
  SOFT_COVER: "SOFT_COVER",
  WOOD_COVER: "WOOD_COVER",
  SLIP_COVER: "SLIP_COVER",
  OTHERS: "OTHERS",
} as const;